+++ src/frontend/components/VideoPlayer/index.tsx
@@ -139,15 +139,14 @@ class BaseVideoPlayer extends React.Component<
           crossOrigin="anonymous"
           poster={thumbnailUrls[720]}
         >
-          {!this.state.isDashSupported &&
-            (Object.keys(video.urls.mp4) as videoSize[]).map(size => (
-              <source
-                key={video.urls.mp4[size]}
-                size={size}
-                src={video.urls.mp4[size]}
-                type="video/mp4"
-              />
-            ))}
+          {(Object.keys(video.urls.mp4) as videoSize[]).map(size => (
+            <source
+              key={video.urls.mp4[size]}
+              size={size}
+              src={video.urls.mp4[size]}
+              type="video/mp4"
+            />
+          ))}
 
           {/* This is a workaround to force plyr to load its tracks list once
           instantiated. Without this, captions are not loaded correctly, at least, on firefox.